fs: btrfs: Crossport open_ctree_fs_info()
open_ctree_fs_info() is the main entry point to open a btrfs.

This version is a simplfied version of __open_ctree_fd() of btrfs-progs,
the main differences are:
- Parameters on how to specify a block device
  Instead of @fd and @path, uboot uses blk_desc and disk_partition_t.

- Remove open_ctree flags
  There will not be multiple open ctree modes in uboot.

Otherwise the function structure are all kept the same.

With open_ctree_fs_info() implemented, also introduce the global
current_fs_info pointer to show the current opened btrfs.

Signed-off-by: Qu Wenruo <wqu@suse.com>